TheGreatEscape.c: searchlight_sub -> searchlight_mask_test.
searchlight_mask_test tests the mask buffer to see if the hero (or maybe the centre tile...) is out of sight. e.g. if he moves behind one of the huts then he is not visible and the searchlight loses track of him.

The countdown before he is lost is in state->searchlight_state, which is a counter rather than a flag (and will need renaming).