increase/decrease_morale: Use an int for the temporary morale value otherwise the checks can overflow.
This happened as the code was ported using byte-sized values to match the Z80's registers but the adjust-then-test pattern in the converted code is not the same as the simultaneous adjust-and-test of the assembly. The assembly equivalent will get away with wrapping around since it produces a carry irrespective.