Fix: Hero should leave solitary when commandant arrives.
I’d altered charevnt_hero_release in the distant past to jump to the start of set_hero_route rather than partway into it as it should. Doing that stops any routes being set if the hero is in solitary.

So turn set_hero_route’s tail into set_hero_route_force and call that instead.