Break tge_setup apart into tge_setup/tge_menu/tge_setup2.
Do this so that the main game thread loop can check the quit flag on every iteration of the menu and game loops. Without this it's impossible to quit the game cleanly.

There are a couple of infinite loops still around which will need addressing. These have been marked with FIXMEs.