Rework sleep handling.
Remove sleeptype_t from the sleep callback and add a new callback 'stamp' to call prior to any timing-sensitive section of code that is paired with a later 'sleep' callback. 'stamp' records the current time. When 'sleep' is called at the end of the section that recorded start time can be used to work out the appropriate sleep period.

Note that 'stamp' calls will nest so we keep a small stack of them. This stack is reset if the game longjmps.

'sleep' is now accepts a time given in T-states.

The individual sleep periods were analysed in the original game and an accurate number of T-states has been selected for the section of code they represent. For example, the main_loop now has a T-state counter of 367,731 based on an average of runs taken from the original game (should be ~9.5fps). Note that this isn't a perfect recreation of the original game's timing. The original slowed down when more characters appeared on screen; we avoid that since it negatively affects gameplay. The zoombox routine however seemed very wrong without the original timing so we now mimic that by adjusting the sleep period proportional to the size of the zoombox area being drawn.

The key scanning loops in the game which would otherwise run continuously now wait for 1/10th of a second inbetween scans.

Sleep periods are scaled in accordance with the game speed when the speed is changed.

The interior_delay_loop routine called from main_loop has been removed as it serves no purpose now the main_loop regulates itself.

This commit does not include any changes to the Windows platform.

This was coded during the music and sound work but is being committed ahead of it.