Get the SDL & Emscripten builds going.
This adds a variation on the generic build which runs the UI and game core in the same thread.

Limitations:
- No audio
- Display doesn't react to changes in window size
- Keys aren't fully mapped yet (just use cursor keys + dot for now)
- Timing is lumpy
- No speed controls
- Border callback is ignored (no issue)

The single threaded setup creates problems when the game wants to sleep, such as during zoomboxing and when choosing keys when starting the game. Starting the game in kempston mode allows it to proceed presently.