Fix: Red flag activating unexpectedly.
There was stray code in `in_permitted_area` in the original game which, while it remained side effect-free in the original, when converted it ended up meddling with route.step. In the most obvious instance this pushed route.step over from 15 to 16 when the hero was wandering which made later code decide that he should have been in the exercise yard. The flag would turn red, the hero would halt his automatic perambulations and promptly get thrown into solitary.

+ Update BUGS.md - numbered bugs.