Fix: Out-of-bounds memory accesses in sprite plotters.
The individual byte plotting commands form their masked pixels irrespective of whether they are going to appear on-screen or not. This is not a problem in the original game but in the conversion it results in out-of-bounds memory accesses.

Adjust the code so that the state->enable_* condition wraps the mask pixel-forming MASK() operation as well as the actual store through screenptr.