Fix: Out-of-bounds memory access in render_mask_buffer.
At the end of a mask being rendered in render_mask_buffer the code will execute the trailing case when the mask buffer pointer is readied for the next row. In the original code this accessed out-of-bounds bytes with no ill effect, however that cannot work in the conversion.

To fix we now test for the final row and skip the trailing code.