Add locking to screen accesses.
Rather than handing out a direct pointer to a shared pixel buffer in the draw callback we now have the app code call a new zxspectrum_claim_screen() function when it's ready to draw the screen. This takes a lock which is released when zxspectrum_release_screen() is called.

Forcing the app to claim at point of redraw, as late as possible, could perhaps reduce the amount of screen buffer conversion that occurs.

To avoid converting an in-progress screen an extra Spectrum screen buffer is added inside the ZXSpectrum layer. This is copied from the real screen in draw callbacks then later used for conversion. (It would be better to avoid this, if possible).

To know what region is due for conversion we now maintain a dirty rectangle in the ZXSpectrum layer. This is the union of all dirty rectangles we're told of up until zxspectrum_claim_screen() does a conversion. This lets us dodge any conversions when the screen hasn't really changed.

Note: This will break the Windows build until the pthread stuff is abstracted away. I will need to create or borrow a library to fix that.