get_greatest_itemstruct: Remove weird assignment to state->IY.
This was confusing since the game only ever uses IY elsewhere as a current vischar pointer, but in this code it’s used as a return value. I’d originally followed the pattern of other IY use and assigned it to state->IY but in this case that’s not correct. The fix is just to remove the assignment.