Check for shutdown when in non-yielding game thread loops.
This adds a return value to the sleep() callback which signals when the game thread is being shut down. Previously, certain points in the game thread would spin forever and not yield to shut down, e.g. when defining keys and waiting for the user to confirm game restart.

In menu code the shut down value is tested for and if found is returned as a reason code to the app side.
In game code the shut down value is tested for and if found causes a squash_stack_goto_main(), resulting in tge_main() returning to the app side.

Two utility menudelay() and gamedelay() functions are added to support this.

As a result of this keyscan_all() now rate limits its key scanning.